# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "스터디 모집·신청 서비스",
    "description": "\n해커톤 연습용 스터디 모집·신청 서비스 API.\n\n**인증 없음**: 요청자는 `user_id`로만 구분한다.\n- `GET` / `DELETE` → `user_id`를 **쿼리**로 전달\n- `POST` / `PATCH` → `user_id`를 **body**로 전달\n- 방장 전용 API는 `user_id ≠ owner_id`이면 403\n\n에러 응답은 모두 `{\"detail\": \"한국어 메시지\"}` 형식이다. 시간은 KST naive datetime(`YYYY-MM-DDTHH:MM:SS`).\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/studies/{study_id}/applications": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "참여 신청",
        "description": "모집중(RECRUITING) 스터디만 신청 가능(아니면 400).\n방장·기존 멤버·대기 중(PENDING) 중복 신청은 409. 거절/취소된 뒤 재신청은 가능.",
        "operationId": "apply_studies__study_id__applications_post",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "중복/충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "신청 목록 (방장)",
        "description": "신청자 닉네임 포함. `status`로 필터 가능.",
        "operationId": "list_applications_studies__study_id__applications_get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "요청자(방장) user_id",
              "title": "User Id"
            },
            "description": "요청자(방장) user_id"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationWithUser"
                  },
                  "title": "Response List Applications Studies  Study Id  Applications Get"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}": {
      "delete": {
        "tags": [
          "applications"
        ],
        "summary": "신청 취소 (본인)",
        "description": "대기 중(PENDING)인 신청만 취소 가능. 상태가 CANCELED로 바뀐다.",
        "operationId": "cancel_application_applications__application_id__delete",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "요청자(신청자 본인) user_id",
              "title": "User Id"
            },
            "description": "요청자(신청자 본인) user_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}/approve": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "신청 승인 (방장)",
        "description": "PENDING만 승인 가능, 정원이 가득 차면 400.\n한 트랜잭션에서 상태 변경 + 멤버 생성 + 정원 도달 시 스터디 CLOSED.",
        "operationId": "approve_applications__application_id__approve_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequesterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}/reject": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "신청 거절 (방장)",
        "description": "PENDING만 거절 가능.",
        "operationId": "reject_applications__application_id__reject_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequesterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "examples": [
              2
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "",
            "examples": [
              "열심히 참여하겠습니다!"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ApplicationCreate"
      },
      "ApplicationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "user_id",
          "message",
          "status",
          "created_at"
        ],
        "title": "ApplicationOut"
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELED"
        ],
        "title": "ApplicationStatus"
      },
      "ApplicationWithUser": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "user_id",
          "message",
          "status",
          "created_at",
          "nickname"
        ],
        "title": "ApplicationWithUser"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "방장만 할 수 있는 작업입니다."
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequesterBody": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "RequesterBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}